LayoutTests/fast/repaint/background-attachment-fixed-in-composited-scroll-expected.txt

Test that a scroll of an overflow scrolling element, with a composoited `background-attachment:fixed` child, repaints

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS layerTreeAsText.indexOf('(rect 0.00 0.00 300.00 750.00)') > -1 is true
PASS successfullyParsed is true

TEST COMPLETE
